chore: Organize schemas.
